// src/app/operations/models/bkmcargo.ts
export class BkmCargo {
  bc_pkid: string;
  bc_parent_id: string;
  bc_desc: string;
  bc_ritc_id: string;
  bc_ritc_code: string;
  bc_ritc_name: string;
  bc_wt: number;
  bc_wt_unit: string;
  bc_cbm: number;
  bc_cbm_unit: string;
  bc_pkg: number;
  bc_pkg_id: string;
  bc_pkg_code: string;
  bc_pkg_name: string;

}